+++ DUI3-DX/Converters/ArcGIS/Speckle.Converters.ArcGIS3.DependencyInjection/AutofacArcGIS3ConverterModule.cs
@@ -18,6 +18,8 @@ public class AutofacArcGISConverterModule : Module
     builder.RegisterType<ArcGISConverterToSpeckle>().As<ISpeckleConverterToSpeckle>().InstancePerLifetimeScope();
     builder.RegisterType<ArcGISConverterToHost>().As<ISpeckleConverterToHost>().InstancePerLifetimeScope();
     builder.RegisterType<FeatureClassUtils>().As<IFeatureClassUtils>().InstancePerLifetimeScope();
+    builder.RegisterType<FieldsUtils>().As<IFieldsUtils>().InstancePerLifetimeScope();
+    builder.RegisterType<OtherUtils>().As<IOtherUtils>().InstancePerLifetimeScope();
     builder.RegisterType<ArcGISProjectUtils>().As<IArcGISProjectUtils>().InstancePerLifetimeScope();
 
     builder
