+++ DUI3-DX/Converters/ArcGIS/Speckle.Converters.ArcGIS3/Utils/IFeatureClassUtils.cs
@@ -15,10 +15,5 @@ public interface IFeatureClassUtils
     IRawConversion<IReadOnlyList<Base>, ACG.Geometry> gisGeometryConverter
   );
   void AddFeaturesToTable(Table newFeatureClass, List<GisFeature> gisFeatures, List<FieldDescription> fields);
-  public string CleanCharacters(string key);
-  public RowBuffer AssignFieldValuesToRow(RowBuffer rowBuffer, List<FieldDescription> fields, GisFeature feat);
-  public object? FieldValueToNativeType(FieldType fieldType, object? value);
-  public List<FieldDescription> GetFieldsFromSpeckleLayer(VectorLayer target);
-  public FieldType GetFieldTypeFromInt(int fieldType);
   public ACG.GeometryType GetLayerGeometryType(VectorLayer target);
 }
