+++ DUI3-DX/Converters/ArcGIS/Speckle.Converters.ArcGIS3/Utils/IOtherUtils.cs
@@ -1,0 +1,6 @@
+namespace Speckle.Converters.ArcGIS3.Utils;
+
+public interface IOtherUtils
+{
+  public string CleanCharacters(string key);
+}
